Restructure entire backend project
