--- chainsaw-main.code-workspace
@@ -1,12 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		}
-	],
-	"settings": {
-		"terminal.integrated.automationProfile.linux": {
-			"path": "/usr/local/bin/bash_vscode_automation",
-		}
-	}
-}
