+++ docker/backend.Dockerfile
@@ -1,6 +1,8 @@
-FROM python:3.9
+FROM python:3.10
 
 WORKDIR /code
+
+RUN python3 -m pip install pipenv
 
 RUN mkdir docker
