+++ server/Pipfile
@@ -1,0 +1,48 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+anyio = "==3.6.1"
+asgiref = "==3.5.2"
+black = "==22.6.0"
+certifi = "==2022.6.15"
+charset-normalizer = "==2.1.0"
+click = "==8.1.3"
+dnspython = "==2.2.1"
+email-validator = "==1.2.1"
+fastapi = "==0.79.0"
+h11 = "==0.13.0"
+httptools = "==0.4.0"
+idna = "==3.3"
+itsdangerous = "==2.1.2"
+jinja2 = "==3.1.2"
+markupsafe = "==2.1.1"
+motor = "==3.0.0"
+mypy-extensions = "==0.4.3"
+orjson = "==3.7.11"
+pathspec = "==0.9.0"
+platformdirs = "==2.5.2"
+pydantic = "==1.9.1"
+pymongo = "==4.2.0"
+python-dotenv = "==0.20.0"
+python-multipart = "==0.0.5"
+pyyaml = "==6.0"
+requests = "==2.28.1"
+six = "==1.16.0"
+sniffio = "==1.2.0"
+starlette = "==0.19.1"
+tomli = "==2.0.1"
+typing-extensions = "==4.3.0"
+ujson = "==5.4.0"
+urllib3 = "==1.26.11"
+uvicorn = "==0.17.6"
+uvloop = "==0.16.0"
+watchgod = "==0.8.2"
+websockets = "==10.3"
+
+[dev-packages]
+
+[requires]
+python_version = "3.10"
