--- server/api/__init__.py
@@ -1,3 +1,0 @@
-from api.theme import Theme
-from api.journal_entry import JournalEntry
-from api.journal_theme_data import JournalThemeData
