+++ server/app/api/__init__.py
@@ -1,0 +1,5 @@
+"""This module contains all the business logic required to make the application function as expected.
+"""
+from . import journal_entry
+from . import journal_theme_data
+from . import theme
