+++ server/app/constants/collection.py
