+++ server/app/helpers/__init__.py
@@ -1,0 +1,3 @@
+"""Module with helpers for functions in app/api/
+"""
+from .journal_entry import get_journal_entry_helper
