+++ server/app/models/__init__.py
@@ -1,0 +1,7 @@
+"""Module with all the models powering this application
+"""
+
+from . import base
+from . import journal_entry
+from . import journal_theme
+from . import journal_theme_data
