+++ server/app/router/__init__.py
@@ -1,0 +1,3 @@
+from app.router.journal_entry import JOURNAL_ENTRY_ROUTER
+from app.router.journal_theme import JOURNAL_THEME_ROUTER
+from app.router.journal_theme_data import JOURNAL_THEME_DATA_ROUTER
