+++ server/app/router/journal_theme_data.py
@@ -1,7 +1,8 @@
 from fastapi import Depends, APIRouter
-from api import JournalThemeData
-from models import journal_theme_data as journal_theme_data_models
-from repository.journal_theme_data import get_journal_theme_data_repo
+
+from app.api import journal_theme_data
+from app.models import journal_theme_data as journal_theme_data_models
+from app.repository.journal_theme_data import get_journal_theme_data_repo
 
 JOURNAL_THEME_DATA_ROUTER = APIRouter()
 
@@ -9,4 +10,4 @@ JOURNAL_THEME_DATA_ROUTER = APIRouter()
 async def get_journal_theme_data(input: journal_theme_data_models.GetJournalThemeDataInput,
     journal_theme_data_repo=Depends(get_journal_theme_data_repo)
 ):
-    return await JournalThemeData(journal_theme_data_repo).get(input)
+    return await journal_theme_data.get()
