--- server/router/__init__.py
@@ -1,3 +1,0 @@
-from router.journal_entry import JOURNAL_ENTRY_ROUTER
-from router.journal_theme import JOURNAL_THEME_ROUTER
-from router.journal_theme_data import JOURNAL_THEME_DATA_ROUTER
