+++ server/scripts/seed_db.py
@@ -1,0 +1,8 @@
+from app.connections.database import _get_client
+
+client = _get_client()
+
+# add themes
+# add theme data
+# add dummy journal entries
+# run this script when starting the docker containers
